+++ src/test/java/com/hashedin/hu/LeaveManagerTest.java
@@ -14,7 +14,7 @@ public class LeaveManagerTest {
     @Test
     public void testValidateLeave() {
         LeaveManager manager = new LeaveManager();
-        Employee e = new Employee("emp1",1,10, Gender.MALE,"2019-01-01");
+        Employee e = new Employee("emp1", (long) 1,10, Gender.MALE,"2019-01-01");
         LeaveRequest leave = new LeaveRequest(e, LocalDate.of(2019,3,1), LocalDate.of(2019,3,4),true);
         LeaveResponse response = manager.ApplyLeave(leave);
         assertEquals(response.status,LeaveStatus.APPROVED);
@@ -22,7 +22,7 @@ public class LeaveManagerTest {
     @Test
     public void testLeaveMoreThanEmpHas() {
         LeaveManager manager = new LeaveManager();
-        Employee e = new Employee("emp1",1,1, Gender.MALE,"2019-01-01");
+        Employee e = new Employee("emp1", (long) 1,1, Gender.MALE,"2019-01-01");
         LeaveRequest leave = new LeaveRequest(e, LocalDate.of(2019,3,1), LocalDate.of(2019,3,10),true);
         LeaveResponse response = manager.ApplyLeave(leave);
         assertEquals(response.status,LeaveStatus.REJECTED);
@@ -31,7 +31,7 @@ public class LeaveManagerTest {
     @Test
     public void checkForWeekDays(){
         LeaveManager manager = new LeaveManager();
-        Employee e = new Employee("emp2",3,20,Gender.FEMALE,"2019-01-01");
+        Employee e = new Employee("emp2", (long) 3,20,Gender.FEMALE,"2019-01-01");
         LeaveRequest leave = new LeaveRequest(e,LocalDate.of(2019,3,1),LocalDate.of(2019,3,4),false);
         LeaveResponse response = manager.ApplyLeave(leave);
         assertEquals(1,e.no_of_leaves_taken);
@@ -39,7 +39,7 @@ public class LeaveManagerTest {
     }
     @Test
     public void testMaternityLeave(){
-        Employee e = new Employee("emp4",5,20,Gender.FEMALE,"2019-01-01");
+        Employee e = new Employee("emp4", (long) 5,20,Gender.FEMALE,"2019-01-01");
         LeaveRequest leave = new LeaveRequest(e,LocalDate.of(2019,7,1),null,true);
         leave.setMaternityLeave();
         LeaveResponse response = manager.ApplyLeave(leave);
@@ -48,7 +48,7 @@ public class LeaveManagerTest {
 
     @Test
     public void testMaternityLeaveForMale(){
-        Employee e = new Employee("emp4",5,20,Gender.MALE,"2019-01-01");
+        Employee e = new Employee("emp4", (long) 5,20,Gender.MALE,"2019-01-01");
         LeaveRequest leave = new LeaveRequest(e,LocalDate.of(2019,3,1),LocalDate.of(2019,4,5),true);
         leave.setRequestedDate(LocalDate.of(2019,1,1));
         leave.setTypes(LeaveTypes.MATERNITY);
@@ -57,7 +57,7 @@ public class LeaveManagerTest {
     }
     @Test
     public void testPaternityLeaveForMale(){
-        Employee e = new Employee("emp4",5,20,Gender.MALE,"2019-01-01");
+        Employee e = new Employee("emp4", (long) 5,20,Gender.MALE,"2019-01-01");
         LeaveRequest leave = new LeaveRequest(e,LocalDate.of(2019,3,1),LocalDate.of(2019,3,5),true);
         leave.setTypes(LeaveTypes.PATERNITY);
         LeaveResponse response = manager.ApplyLeave(leave);
@@ -65,7 +65,7 @@ public class LeaveManagerTest {
     }
     @Test
     public void testPaternityLeaveForFemale(){
-        Employee e = new Employee("emp4",5,20,Gender.FEMALE,"2019-01-01");
+        Employee e = new Employee("emp4", (long) 5,20,Gender.FEMALE,"2019-01-01");
         LeaveRequest leave = new LeaveRequest(e,LocalDate.of(2019,3,1),LocalDate.of(2019,3,5),true);
         leave.setTypes(LeaveTypes.PATERNITY);
         LeaveResponse response = manager.ApplyLeave(leave);
@@ -74,7 +74,7 @@ public class LeaveManagerTest {
     @Test
     public void testValidateLeaveDuplicate() {
         LeaveManager manager = new LeaveManager();
-        Employee e = new Employee("emp1",1,10, Gender.MALE,"2019-01-01");
+        Employee e = new Employee("emp1", (long) 1,10, Gender.MALE,"2019-01-01");
         LeaveRequest leave = new LeaveRequest(e, LocalDate.of(2019,3,1), LocalDate.of(2019,3,4),true);
         LeaveRequest leave2 = new LeaveRequest(e, LocalDate.of(2019,3,2), LocalDate.of(2019,3,3),true);
         LeaveResponse response = manager.ApplyLeave(leave);
@@ -83,7 +83,7 @@ public class LeaveManagerTest {
     }
 @Test
     public void testApplyLeaveWithoutCompoff(){
-    Employee e = new Employee("emp1",1,10, Gender.MALE,"2019-01-01");
+    Employee e = new Employee("emp1", (long) 1,10, Gender.MALE,"2019-01-01");
     LeaveRequest leave = new LeaveRequest(e,LocalDate.of(2019,3,1),LocalDate.of(2019,3,2),false);
     leave.setTypes(LeaveTypes.COMP_OFF);
     LeaveResponse response = manager.ApplyLeave(leave);
@@ -91,7 +91,7 @@ public class LeaveManagerTest {
 }
 @Test
     public void testAddCompOffAndApply(){
-    Employee e = new Employee("emp1",1,10, Gender.MALE,"2019-01-01");
+    Employee e = new Employee("emp1", (long) 1,10, Gender.MALE,"2019-01-01");
     e.addCompoff(LocalDateTime.of(2019, Month.FEBRUARY,16,9,00),
             LocalDateTime.of(2019,Month.FEBRUARY,16,20,00));
     LeaveRequest leave = new LeaveRequest(e,LocalDate.of(2019,3,1),LocalDate.of(2019,3,2),false);
@@ -102,7 +102,7 @@ public class LeaveManagerTest {
 @Test
     public void testAddCompoffAndRedeemAfterMonth()
 {
-    Employee e = new Employee("emp1",1,10, Gender.MALE,"2019-01-01");
+    Employee e = new Employee("emp1", (long) 1,10, Gender.MALE,"2019-01-01");
     e.addCompoff(LocalDateTime.of(2019, Month.FEBRUARY,16,9,00),
             LocalDateTime.of(2019,Month.FEBRUARY,16,18,00));
     LeaveRequest leave = new LeaveRequest(e,LocalDate.of(2019,4,1),LocalDate.of(2019,4,2),false);
@@ -112,20 +112,20 @@ public class LeaveManagerTest {
 }
 @Test(expected = IllegalArgumentException.class)
     public void testInvalidDates() {
-    Employee e = new Employee("emp1", 1, 10, Gender.MALE, "2019-01-01");
+    Employee e = new Employee("emp1", (long) 1, 10, Gender.MALE, "2019-01-01");
     LeaveRequest leave = new LeaveRequest(e, LocalDate.of(2019, 3, 10), LocalDate.of(2019, 3, 1), false);
     LeaveResponse response = manager.ApplyLeave(leave);
     }
     @Test(expected = IllegalArgumentException.class)
     public void testApplyLeaveForPastDate(){
-        Employee e = new Employee("emp1", 1, 10, Gender.MALE, "2019-01-01");
+        Employee e = new Employee("emp1", (long) 1, 10, Gender.MALE, "2019-01-01");
         LeaveRequest leave = new LeaveRequest(e, LocalDate.of(2019, 2, 10), LocalDate.of(2019, 2, 11), false);
         LeaveResponse response= manager.ApplyLeave(leave);
     }
     @Test
     public void testForOpitonalLeave()
     {
-        Employee e = new Employee("emp1", 1, 0, Gender.MALE, "2019-01-01");
+        Employee e = new Employee("emp1", (long) 1, 0, Gender.MALE, "2019-01-01");
         LeaveRequest leave = new LeaveRequest(e,LocalDate.of(2019,8,11),LocalDate.of(2019,8,13),false);
         LeaveResponse response = manager.ApplyLeave(leave);
         assertEquals(1,leave.empoyee.no_of_leaves_taken);
@@ -133,7 +133,7 @@ public class LeaveManagerTest {
     @Test
     public void testEmployeeIsApplyingForTwoOptionalLeaves()
     {
-        Employee e = new Employee("emp1", 1, 0, Gender.MALE, "2019-01-01");
+        Employee e = new Employee("emp1", (long) 1, 0, Gender.MALE, "2019-01-01");
         LeaveRequest leave = new LeaveRequest(e,LocalDate.of(2019,8,11),LocalDate.of(2019,8,13),false);
         LeaveResponse response = manager.ApplyLeave(leave);
         LeaveRequest leave2 = new LeaveRequest(e,LocalDate.of(2019,3,19),LocalDate.of(2019,3,19),false);
@@ -143,7 +143,7 @@ public class LeaveManagerTest {
     @Test
     public void tesApplyForSabbatical()
     {
-        Employee e = new Employee("emp1", 1, 0, Gender.MALE, "2019-01-01");
+        Employee e = new Employee("emp1", (long) 1, 0, Gender.MALE, "2019-01-01");
         LeaveRequest leave = new LeaveRequest(e,LocalDate.of(2021,3,1),LocalDate.of(2021,5,1),false);
         leave.types = LeaveTypes.SABBATICAL;
         LeaveResponse response = manager.ApplyLeave(leave);
@@ -152,7 +152,7 @@ public class LeaveManagerTest {
     @Test
     public void testApplyForSabbaticalWothout2YearsOfExperience()
     {
-        Employee e = new Employee("emp1", 1, 0, Gender.MALE, "2019-01-01");
+        Employee e = new Employee("emp1", (long) 1, 0, Gender.MALE, "2019-01-01");
         LeaveRequest leave = new LeaveRequest(e,LocalDate.of(2019,3,1),LocalDate.of(2019,5,1),false);
         leave.types = LeaveTypes.SABBATICAL;
         LeaveResponse response = manager.ApplyLeave(leave);
